requirements.txt
flask
twilio
python-dotenv
datefinder
pyngrok
google-generativeai
langchain
PyPDF2
#chromadb
#faiss-cpu
langchain_google_genai
langchain-community
Pillow 
pytesseract
opencv-python
mysqlclient
datefinder
PyPDF2
pdfplumber
reportlab
ocrmypdf
pandas